feat(spring-core): 增加单例bean模式的支持，并规范bean的命名规则，保证全局唯一
