+++ go.mod
@@ -7,4 +7,7 @@ require (
 	github.com/labstack/echo v3.3.10+incompatible
 	github.com/labstack/gommon v0.3.0 // indirect
 	github.com/spf13/cast v1.3.0
+	github.com/stretchr/testify v1.4.0
 )
+
+go 1.13
