+++ spring-core/testdata/bar/bar.go
@@ -1,0 +1,4 @@
+package bar
+
+type Demo struct {
+}
