+++ spring-core/testdata/foo/foo.go
@@ -1,0 +1,4 @@
+package foo
+
+type Demo struct {
+}
